% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connect_acd}
\alias{connect_acd}
\title{Connect to Tassolog's ACD database}
\usage{
connect_acd()
}
\description{
Connect to the "acd" database on Tassolog. A [acd] RMariaDB connection should be defined in your my.cnf file.
}
\keyword{acd}
\keyword{connect}
\keyword{connect_acd()}
