% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connect_agents_analytics}
\alias{connect_agents_analytics}
\title{Connect to Tassolog's agents_analytics database}
\usage{
connect_agents_analytics()
}
\description{
Connect to the "agents_analytics" database on Tassolog. A [agents_analytics] RMariaDB connection should be defined in your my.cnf file.
}
\keyword{agents}
\keyword{analytics}
\keyword{connect}
\keyword{connect_agents_analytics()}
