% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connect_allen_release}
\alias{connect_allen_release}
\title{Connect to Tassolog's allen_release database}
\usage{
connect_allen_release()
}
\description{
Connect to the "allen_release" database on Tassolog. A [allen_release] RMariaDB connection should be defined in your my.cnf file.
}
\keyword{allen}
\keyword{connect}
\keyword{connect_allen_release()}
\keyword{release}
