% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connect_big3}
\alias{connect_big3}
\title{Connect to Tassolog's allen_release, reporting, and urd databases.}
\usage{
connect_big3()
}
\description{
A convenience function. This invokes the connect_allen_release(), connect_reporting(), and connect_urd() functions.
}
\keyword{allen}
\keyword{connect}
\keyword{connect_big3()}
\keyword{release}
\keyword{reporting}
\keyword{urd}
