% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connect_billing}
\alias{connect_billing}
\title{Connect to Tassolog's billing database}
\usage{
connect_billing()
}
\description{
Connect to the "billing" database on Tassolog. A [billing] RMariaDB connection should be defined in your my.cnf file.
}
\keyword{billing}
\keyword{connect}
\keyword{connect_billing()}
