% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connect_grouphead}
\alias{connect_grouphead}
\title{Connect to Tassolog's grouphead database}
\usage{
connect_grouphead()
}
\description{
Connect to the "grouphead" database on Tassolog. A [grouphead] RMariaDB connection should be defined in your my.cnf file.
}
\keyword{connect}
\keyword{connect_grouphead()}
\keyword{grouphead}
