% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connect_reporting}
\alias{connect_reporting}
\title{Connect to Tassolog's reporting database}
\usage{
connect_reporting()
}
\description{
Connect to the "reporting" database on Tassolog. A [reporting] RMariaDB connection should be defined in your my.cnf file.
}
\keyword{connect}
\keyword{connect_reporting()}
\keyword{reporting}
