% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connect_snowflake}
\alias{connect_snowflake}
\title{Connect to Snowflake Data Warehouse}
\usage{
connect_snowflake()
}
\description{
Connect to Snowflake Data Warehouse, our data warehouse. A [convodatawarehouse] odbc connection should be defined in your .odbc.ini file.
}
\keyword{connect}
\keyword{connect_snowflake()}
\keyword{snowflake}
