% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connect_urd}
\alias{connect_urd}
\title{Connect to Tassolog's urd database}
\usage{
connect_urd()
}
\description{
Connect to the "urd" database on Tassolog. A [urd] RMariaDB connection should be defined in your my.cnf file.
}
\keyword{connect}
\keyword{connect_urd()}
\keyword{urd}
