% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convor-package.R
\docType{package}
\name{convor-package}
\alias{convor}
\alias{convor-package}
\title{convor: R Functions for Convo Data Science}
\description{
A set of functions to make it easier to work with Convo's databases as well as some common operations for certain datasets.
}
\author{
\strong{Maintainer}: Adam Stone \email{adam.stone@convorelay.com}

}
\keyword{internal}
