% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{get_live_tdns}
\alias{get_live_tdns}
\title{Get Live TDNs On a Specific Date}
\usage{
get_live_tdns(x)
}
\arguments{
\item{x}{Date value}
}
\description{
For a specified date, this function will return a list of all TDNs from locations created up to that date, and all archived TDNs that were created prior to that date and deleted after that date. Also will cast phone numbers as character type, and remove duplicate TDNs. Requires an active connection to allen_release via connect_allen_release().
}
\examples{
get_live_tdns('2018-01-01') # Pulls all live TDNs on January 1, 2018.
}
\keyword{active}
\keyword{live}
\keyword{tdn}
\keyword{user}
