% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{get_urd_regs}
\alias{get_urd_regs}
\title{Retrieve URD Registration Data}
\usage{
get_urd_regs()
}
\description{
Often we need to pull this information from our URD database. This will produce a table of each TDN's latest URD registration data and status.
}
\examples{
get_urd_regs()
}
\keyword{reg}
\keyword{registration}
\keyword{tdn}
\keyword{urd}
