% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{load_urd_users}
\alias{load_urd_users}
\title{Load URD Users file}
\usage{
load_urd_users()
}
\description{
Will read in the newest Users-M-DD-YYYY.xlsx file downloaded from RoulkaLoube's URD portal, and clean up dates/fields. You must have downloaded the file already and it must be in the `~/Downloads` folder.
}
\examples{
load_urd_users()
load_urd_users()
}
\keyword{rl}
\keyword{rla}
\keyword{urd}
\keyword{xls}
\keyword{xlsx}
