% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{tbl_snowflake}
\alias{tbl_snowflake}
\title{Retrieve results from a query to Snowflake}
\usage{
tbl_snowflake(x)
}
\description{
Will send the query to Snowflake and retrieve results in tidyverse-friendly format (i.e., a tibble with lower-case names).
}
\keyword{query}
\keyword{snowflake}
\keyword{tbl_snowflake()}
