% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{write_csv_rla}
\alias{write_csv_rla}
\title{Generate CSV Files in RLA-Friendly Format}
\usage{
write_csv_rla(df, filename)
}
\arguments{
\item{df}{Data fame to be written to.}

\item{filename}{File path and filename to write file to.}
}
\description{
Generates CSV files with CRLF line endings, `""` for `NA` values, and no row numbers. Intended for files to be uploaded to RLA. .txt or .csv endings may be used.
}
\examples{
write_csv_rla(billing_df, "~/Desktop/billing_df.txt")
write_csv_rla(df, filename)
}
\keyword{csv}
\keyword{file}
\keyword{rl}
\keyword{rla}
\keyword{text}
