% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convor-package.R
\docType{package}
\name{convor-package}
\alias{convor}
\alias{convor-package}
\title{convor: R Functions for Convo Data Science}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

A set of functions to make it easier to work with Convo's databases as well as some common operations for certain datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/foundinblank/convor}
  \item Report bugs at \url{https://github.com/foundinblank/convor/issues}
}

}
\author{
\strong{Maintainer}: Adam Stone \email{adam.stone@convorelay.com}

}
\keyword{internal}
