% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{parse_phone_number}
\alias{parse_phone_number}
\title{Parse Phone Numbers}
\usage{
parse_phone_number(x)
}
\arguments{
\item{x}{Character vector of values to parse.}
}
\description{
Drops all non-numeric characters from phone numbers (and maybe any other strings).
}
\examples{
x <- c('(858) 555-1234', '(619) 555-6898', '585-555-5432')
parse_phone_number(x)
}
\keyword{number}
\keyword{parse}
